account 기능 수정
계정 기능 수정

+++ metanet_internship_4team/src/main/java/com/metanet/repository/LikesRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import com.metanet.domain.Likes;
 
 @Repository
-public interface LikesRepository extends JpaRepository< Likes, Long>
+public interface LikesRepository extends JpaRepository< Likes, Integer>
 {
